pyproject.toml
[tool.poetry]
name = "dev-agent"
version = "1.0.0"
description = "Autonomous Developer Assistant Platform"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
slowapi = "^0.1.9"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.29.0"
anthropic = "^0.34.2"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
ruff = "^0.6.0"
black = "^24.0.0"
mypy = "^1.11.0"

# ── Ruff (linter) ──────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in default arguments (FastAPI Depends pattern)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]   # allow assert in tests

# ── Black (formatter) ──────────────────────────────────────────────────────
[tool.black]
line-length = 100
target-version = ["py312"]

# ── Mypy (type checker) ────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["tests/"]

# ── Pytest ─────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
